isolate RFC6979 for testing
